Added DebugEcho... now knows about LINENO for emacs Major cleanup Now only does s_bootstrap if the xyz.vmax is missing Only builds the cmap if the cmap is missing Added 8 bit component volumes and iv files for viewing

+++ bootvol-thesis.bash
@@ -26,19 +26,74 @@
 # visualization of it.  Hopefully later, I will get to processing the
 # Owens Lake data.
 
+
+
+######################################################################
+# Debugging
+######################################################################
+
+declare -ri EXIT_FAILURE=1
+declare -ri EXIT_SUCCESS=0
+
+
+declare -ri TERSE=1
+declare -ri TRACE=4
+declare -ri VERBOSE=8
+declare -ri BOMBASTIC=16
+
+if [ -z "$VERBOSITY" ]; then
+    declare -i debugLevel=4
+else
+    declare -i debugLevel=$VERBOSITY
+fi
+
+# Twisted way to get down to the fundamental script name.
+tmp=${0##/*/}
+tmp=${tmp%%.bash}
+tmp=${tmp##*.}
+tmp=${tmp##*/}
+declare -r SCRIPT_NAME=$tmp
+
+# $1 is the level to compare against debugLevel
+# $2 is line number
+# $3 is the string to echo to stdout.
+DebugEcho()
+{
+    declare -ir val=$1
+    if [ "$debugLevel" -ge "$val" ]; then
+	#echo $2
+	echo "${SCRIPT_NAME}.bash:$2: (`printf "%02d" $1`) $3"
+    fi
+}
+
+#DebugEcho $TERSE     "Terse is on"
+#DebugEcho $TRACE     "Trace is on"
+DebugEcho $VERBOSE    $LINENO  "Verbose is on"
+DebugEcho $BOMBASTIC  $LINENO  "Bombastic is on"
+
+DebugEcho $TERSE $LINENO "debugLevel           = $debugLevel"
+
+######################################################################
+# Local tuning
+######################################################################
+
+
+
 export PATH=${PATH}:.:..
-cells=100
-draw=1000000
+declare -ir cells=100
+declare -ir draw=25000
 declare -ar groups=( as1-crypt as2-slump as3-undef )
 declare -r w=0.5
 declare -r boundaries="-x -${w} -X ${w} -y -${w} -Y ${w} -z -${w} -Z ${w}"
-declare -r debug_level=2
+#declare -r debug_level=11
 
-echo $boundaries
+DebugEcho $TERSE $LINENO "Cells = $cells   Draw = $draw "
+DebugEcho $VERBOSE $LINENO $boundaries
 
-# FIX: if exists Makefile, do this
-make targets-no-test
-#make s_bootstrap xyzdensity xyzvol_cmp volhdr_edit vol_iv volmakecmap vol2vol
+if [ -e Makefile ]; then
+    make targets-no-test
+    #make s_bootstrap xyzdensity xyzvol_cmp volhdr_edit vol_iv volmakecmap vol2vol
+fi
 
 #
 # Bootstrap each of the groups to produce Vmin (V3), Vint(V2), Vmax(V1) volumes
@@ -46,31 +101,33 @@ make targets-no-test
 if [ 1 == 1 ]; then
     for group in "${groups[@]}"; do
 
-	echo Processing $group
-	s_bootstrap ${group}.s -f xyz  -n 3 --out ${group}.xyz. -p --draw ${draw}
-	mv ${group}.xyz.1.vmax ${group}-boot.xyz.vmax
-	mv ${group}.xyz.2.vint ${group}-boot.xyz.vint
-	mv ${group}.xyz.3.vmin ${group}-boot.xyz.vmin
+	DebugEcho $TERSE $LINENO "Processing $group"
 
-	echo Densifying
-	args="  --bpv=16 -w ${cells} -t ${cells} -d ${cells} --verbosity=$debug_level"
-	xyzdensity ${group}-boot.xyz.vmax --out=${group}-vmax.vol -p 1 $args $boundaries
-	echo -n 2
-	xyzdensity ${group}-boot.xyz.vint --out=${group}-vint.vol -p 1 $args $boundaries
-	echo -n 3
-	xyzdensity ${group}-boot.xyz.vmin --out=${group}-vmin.vol -p 1 $args $boundaries
+	if [ ! -e ${group}-boot.xyz.vmax ]; then 
+	    s_bootstrap ${group}.s -f xyz  -n 3 --out ${group}.xyz. -p --draw ${draw} -v $debugLevel
+	    mv ${group}.xyz.1.vmax ${group}-boot.xyz.vmax
+	    mv ${group}.xyz.2.vint ${group}-boot.xyz.vint
+	    mv ${group}.xyz.3.vmin ${group}-boot.xyz.vmin
+	fi
 
-	echo -n 4
+	DebugEcho $TRACE $LINENO  Densifying
+
+	args=" -p 1  --bpv=16 -w ${cells} -t ${cells} -d ${cells} --verbosity=$debugLevel"
+	
+	xyzdensity ${group}-boot.xyz.vmax --out=${group}-vmax.vol $args $boundaries
+	xyzdensity ${group}-boot.xyz.vint --out=${group}-vint.vol $args $boundaries
+	xyzdensity ${group}-boot.xyz.vmin --out=${group}-vmin.vol $args $boundaries
 	xyzdensity --out=${group}-all-1.0.vol -p 1 -b 8 -w ${cells} -t ${cells} -d ${cells} $boundaries \
 	    ${group}-boot.xyz.vmax \
 	    ${group}-boot.xyz.vint \
 	    ${group}-boot.xyz.vmin 
-
 	
 	#volhdr_edit ${group}-all-1.0.vol --out=${group}-all-0.5.vol --xscale=0.5 --yscale=0.5 --zscale=0.5
 	#volhdr_edit ${group}-all-1.0.vol --out=${group}-all-2.0.vol --xscale=2.0 --yscale=2.0 --zscale=2.0
 
-	volmakecmap --cpt=rgba.cpt -o current.cmap --zero=0
+	if [ ! -e current.cmap ]; then 
+	    volmakecmap --cpt=rgba.cpt -o current.cmap --zero=0
+	fi
 
 	vol_iv -b=2 -c ALPHA_BLENDING --numslicescontrol=ALL -p NONE -C current.cmap -o ${group}-all-1.0.iv ${group}-all-1.0.vol
 
@@ -80,12 +137,17 @@ if [ 1 == 1 ]; then
 	volhdr_edit ${group}-vmin.vol --out=tmp $scale && /bin/mv tmp ${group}-vmin.vol
 
 	#
-	# Make each component viewable
+	# Make each component viewable... scale them to min max
 	#
 
-	vol2vol -o ${group}-vmax-8.vol ${group}-vmax.vol  -p 1 --bpv=8 -j 0.5 -k 0.5 -l 0.5
-	vol2vol -o ${group}-vint-8.vol ${group}-vint.vol  -p 1 --bpv=8 -j 0.5 -k 0.5 -l 0.5
-	vol2vol -o ${group}-vmin-8.vol ${group}-vmin.vol  -p 1 --bpv=8 -j 0.5 -k 0.5 -l 0.5
+	vol2vol -o ${group}-vmax-8.vol ${group}-vmax.vol  -p 0 --bpv=8 -j 0.5 -k 0.5 -l 0.5
+	vol2vol -o ${group}-vint-8.vol ${group}-vint.vol  -p 0 --bpv=8 -j 0.5 -k 0.5 -l 0.5
+	vol2vol -o ${group}-vmin-8.vol ${group}-vmin.vol  -p 0 --bpv=8 -j 0.5 -k 0.5 -l 0.5
+
+	ivargs="-b=2 -c ALPHA_BLENDING --numslicescontrol=ALL -p NONE -C current.cmap"
+	vol_iv -o ${group}-vmax-8.iv ${group}-vmax-8.vol $ivargs
+	vol_iv -o ${group}-vint-8.iv ${group}-vint-8.vol $ivargs
+	vol_iv -o ${group}-vmin-8.iv ${group}-vmin-8.vol $ivargs
 
     done
 fi
@@ -95,7 +157,7 @@ fi
 # convert them all to xyz files
 for group in "${groups[@]}"; do
     s_eigs < $group.s > $group.eigs
-    ./eigs2xyz.py $group.eigs > $group.xyz
+    eigs2xyz.py $group.eigs > $group.xyz
     awk '{print $1,$2,$3}' $group.xyz > $group-vmin.xyz
     awk '{print $4,$5,$6}' $group.xyz > $group-vint.xyz
     awk '{print $7,$8,$9}' $group.xyz > $group-vmax.xyz
@@ -104,7 +166,7 @@ done
 for eig_type in vmin vint vmax; do
     echo $eig_type comparing to vol
     for file in as?-?????-${eig_type}.vol; do 
-	./xyzvol_cmp -v $debug_level -d $file as?-?????-$eig_type.xyz > ${eig_type}-${file}.cmp
+	xyzvol_cmp -v $debugLevel -d $file as?-?????-$eig_type.xyz -o ${eig_type}-${file}.cmp
     done
 done
